+++ django_leek/server.py
@@ -3,7 +3,6 @@ import socketserver
 import multiprocessing
 
 from .helpers import load_task
-from . import worker
 from . import helpers
 import django
 
@@ -13,10 +12,37 @@ log = logging.getLogger(__name__)
 Dcommands = {}
 
 
+def target(queue):
+    django.setup()
+    log.info('Worker Starts')
+    done = False
+    while not done:
+        try:
+            task = queue.get()
+            if task is None:
+                done = True
+                break
+                
+            log.info('running task...')
+
+            # Force this forked process to create its own db connection
+            django.db.connection.close()
+
+            task()
+            # workaround to solve problems with django + psycopg2
+            # solution found here: https://stackoverflow.com/a/36580629/10385696
+            django.db.connection.close()
+            log.info('...successfully')
+        except Exception as e:
+            log.exception("...task failed")
+
+    log.info('Worker stopped')
+
+
 class Pool(object):
     def __init__(self):
         self.queue = multiprocessing.Queue()
-        self.worker = multiprocessing.Process(target=worker.target, args=(self.queue,))
+        self.worker = multiprocessing.Process(target=target, args=(self.queue,))
 
 
 class TaskSocketServer(socketserver.BaseRequestHandler):
